feature: add `HasWilaya` and `HasCommune` Traits
